\name{aspics}
\alias{aspics}
\title{aspics list}
\arguments{
  \item{object}{which can be missing, a list of aspic or a
  sigle aspic object}
}
\value{
  aspics object

  aspics object
}
\description{
  A class that contains a list of ASPIC biomass dynamic
  stock assessment model classes

  A class that contains a list of ASPIC biomass dynamic
  stock assessment model classes
}
\examples{
\dontrun{aspics()}
\dontrun{aspics()}
}

