\name{jk}
\alias{jk}
\title{The length of a string (in characters).}
\arguments{
  \item{An}{aspic object}
}
\value{
  An aspic object with parameter estimate
}
\description{
  The length of a string (in characters).
}
\examples{
seq(10)
}
\seealso{
  \code{\link{biodyn}}

  \code{\link{boot}}

  \code{\link{fit}}
}

